+++ packages/core/src/session/index.ts
@@ -1,5 +1,6 @@
 export * from "./git.js";
 export * from "./json.js";
+export * from "./posture.js";
 export * from "./resume-conditions.js";
 export * from "./ritual-sentinel.js";
 export * from "./session-start.js";